Move ExecuteResult to interpreter-api
Aiming at using it from there
